Move some skipped tests to weak-named test project
Now that we have a weak-named project, we can enable some tests that
verify whether proxy types end up in the right module (weak vs strong
named).
